4.23.2017 screenshot

+++ src/main.py
@@ -19,4 +19,4 @@ def double():
 
 
 if __name__ == '__main__':
-    single()
+    double()
